perf(home): implement React lazy loading for below-the-fold components
