+++ Client/src/pages/Home/landingPage.jsx
@@ -1,396 +1,27 @@
-import { useState } from "react";
-import { Link } from "react-router-dom";
-import { useSelector } from "react-redux";
-import {
-  Truck,
-  ShieldCheck,
-  RefreshCcw,
-  Headphones,
-  ShoppingCart,
-  Star,
-  Loader2,
-} from "lucide-react";
-import toast from "react-hot-toast";
-import { useGetProductsQuery } from "../../features/products/store/productsApi";
-import { useAddToCartMutation } from "../../features/cart/store/cartApi";
-import Footer from "../../components/Footer/footer";
-import { optimizeCloudinaryUrl } from "../../utils/productUtils";
+import { lazy, Suspense } from "react";
 
-import landingImage from "../../assets/Home/doctor.webp";
-import FirstImg from "../../assets/Home/first.webp";
-import SecondImg from "../../assets/Home/second.webp";
-import ThirdImg from "../../assets/Home/third.webp";
-import OfferOne from "../../assets/Home/offer_1.webp";
-import OfferTwo from "../../assets/Home/offer_2.webp";
-import Abdo from "../../assets/Home/abdo.webp";
-import Selim from "../../assets/Home/selim.jpeg";
-import Three from "../../assets/Home/three.webp";
-
-const faqs = [
-  {
-    question: "What is your return policy?",
-    answer:
-      "You can return any item within 30 days of purchase for a full refund.",
-  },
-  {
-    question: "Do you offer international shipping?",
-    answer:
-      "Yes, we ship worldwide. Shipping fees depend on the destination country.",
-  },
-  {
-    question: "How can I track my order?",
-    answer:
-      "Once your order is shipped, we will send you a tracking number via email.",
-  },
-];
+import HeroSection from "./components/HeroSection";
+import FeatureCards from "./components/FeatureCards";
+import globalLoader from "../../components/ui/GlobalLoader";
+const PromoBanners = lazy(() => import("./components/PromoBanners"));
+const TopProducts = lazy(() => import("./components/TopProducts"));
+const SpecialOffers = lazy(() => import("./components/SpecialOffers"));
+const Testimonials = lazy(() => import("./components/Testimonials"));
+const FAQAccordion = lazy(() => import("./components/FAQAccordion"));
 
 export default function LandingPage() {
-  const { searchTerm } = useSelector((state) => state.ui);
-  const { isAuthenticated } = useSelector((state) => state.auth);
-  const [activeIndex, setActiveIndex] = useState(null);
-  const [addToCart] = useAddToCartMutation();
-  const [addingProductId, setAddingProductId] = useState(null);
-
-  const queryParams = searchTerm
-    ? { limit: 20, search: searchTerm }
-    : { limit: 4, top_selling: true };
-
-  const { data: productsData, isLoading } = useGetProductsQuery(queryParams);
-  const productsToShow = productsData?.data || [];
-
-  const toggleFAQ = (index) =>
-    setActiveIndex(activeIndex === index ? null : index);
-
-  const handleAddToCart = async (product) => {
-    if (!isAuthenticated) {
-      toast.error("Please login to add items");
-      return;
-    }
-    setAddingProductId(product._id);
-    try {
-      await addToCart({ productId: product._id, quantity: 1 }).unwrap();
-      toast.success("Added to cart successfully");
-    } catch (err) {
-      toast.error("Failed to add to cart", err);
-    } finally {
-      setAddingProductId(null);
-    }
-  };
-
   return (
     <div className="w-full bg-gray-50 dark:bg-[#121212] transition-colors duration-300">
-      <div className="relative w-full overflow-hidden bg-gradient-to-l from-[#0a3c2f] to-[#2c6e49] text-white pt-24 ">
-        <div className="w-full px-4 sm:px-8 lg:px-12">
-          <div className="flex flex-col lg:flex-row items-center justify-between gap-12">
-            <div className="w-full lg:w-1/2 flex flex-col justify-center text-center lg:text-left z-10 pb-12 lg:pb-20">
-              <h1 className="text-4xl md:text-5xl lg:text-6xl font-bold mb-6 leading-tight">
-                Welcome to Rochetta
-              </h1>
-              <p className="text-lg md:text-xl text-green-50 mb-8 max-w-2xl mx-auto lg:mx-0">
-                Your online pharmacy — delivering trusted medicines and care,
-                anytime, anywhere.
-              </p>
-              <div className="flex flex-col sm:flex-row gap-4 justify-center lg:justify-start">
-                <Link
-                  to="/category/pain-relief"
-                  className="px-8 py-3 bg-green-500 hover:bg-green-400 text-white font-bold rounded-xl transition-all shadow-lg active:scale-95"
-                >
-                  Get Started
-                </Link>
-                <button className="px-8 py-3 bg-transparent border-2 border-green-400 hover:bg-green-400/20 text-white font-bold rounded-xl transition-all active:scale-95">
-                  Learn More
-                </button>
-              </div>
-            </div>
-            <div className="w-full lg:w-1/2 flex justify-center lg:justify-end items-end relative top-1">
-              <div className="absolute inset-0 bg-green-400/20 rounded-full blur-3xl scale-150 -z-10"></div>
-              <img
-                src={landingImage}
-                alt="Doctor"
-                width="417"
-                height="625"
-                fetchPriority="high"
-                className="max-h-[450px] lg:max-h-[600px] w-auto object-contain drop-shadow-2xl"
-                style={{ animation: "float 4s ease-in-out infinite" }}
-              />
-            </div>
-          </div>
-        </div>
-      </div>
+      <HeroSection />
+      <FeatureCards />
 
-      <div className="w-full px-4 sm:px-8 lg:px-12 py-16">
-        <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-8">
-          {[
-            {
-              icon: Truck,
-              title: "Free Shipping",
-              desc: "For orders over $199.00",
-            },
-            {
-              icon: ShieldCheck,
-              title: "Secure Payment",
-              desc: "100% secure payment",
-            },
-            {
-              icon: RefreshCcw,
-              title: "Money Back",
-              desc: "30 days return policy",
-            },
-            {
-              icon: Headphones,
-              title: "24/7 Support",
-              desc: "Friendly customer support",
-            },
-          ].map((feature, idx) => (
-            <div
-              key={idx}
-              className="flex flex-col items-center text-center p-6 bg-white dark:bg-[#1e1e1e] rounded-2xl shadow-sm border border-gray-100 dark:border-gray-800 hover:-translate-y-2 transition-transform duration-300"
-            >
-              <div className="w-16 h-16 bg-green-50 dark:bg-green-900/20 rounded-full flex items-center justify-center mb-4">
-                <feature.icon
-                  className="w-8 h-8 text-green-600 dark:text-green-500"
-                  aria-hidden="true"
-                />
-              </div>
-              <h4 className="text-xl font-bold text-gray-900 dark:text-white mb-2">
-                {feature.title}
-              </h4>
-              <p className="text-gray-500 dark:text-gray-400">{feature.desc}</p>
-            </div>
-          ))}
-        </div>
-      </div>
-
-      <div className="w-full px-4 sm:px-8 lg:px-12 py-12">
-        <div className="grid grid-cols-1 md:grid-cols-3 gap-6">
-          {[FirstImg, SecondImg, ThirdImg].map((img, idx) => (
-            <div
-              key={idx}
-              className="overflow-hidden rounded-2xl shadow-md group cursor-pointer"
-            >
-              <img
-                src={img}
-                alt={`Promo ${idx + 1}`}
-                loading="lazy"
-                decoding="async"
-                className="w-full h-full object-cover transform group-hover:scale-105 transition-transform duration-500"
-              />
-            </div>
-          ))}
-        </div>
-      </div>
-
-      <div className="w-full px-4 sm:px-8 lg:px-12 py-16">
-        <h3 className="text-3xl font-bold text-center text-gray-900 dark:text-white mb-10">
-          {searchTerm
-            ? `Search Results for "${searchTerm}"`
-            : "Top Selling Products"}
-        </h3>
-
-        {isLoading ? (
-          <div className="text-center py-10 text-green-600 font-bold flex justify-center items-center gap-2">
-            <RefreshCcw className="w-6 h-6 animate-spin" /> Loading products...
-          </div>
-        ) : (
-          <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6 xl:gap-8">
-            {productsToShow.length > 0 ? (
-              productsToShow.map((product) => (
-                <div
-                  key={product._id}
-                  className="bg-white dark:bg-[#1e1e1e] p-5 rounded-2xl shadow-sm border border-gray-100 dark:border-gray-800 hover:shadow-xl transition-all duration-300 flex flex-col h-full"
-                >
-                  <Link
-                    to={`/product/${product._id}`}
-                    className="block h-48 mb-4 overflow-hidden rounded-xl bg-white  dark:border-gray-700p-2"
-                  >
-                    <img
-                      src={
-                        optimizeCloudinaryUrl(product.image) ||
-                        "/placeholder.png"
-                      }
-                      alt={product.name}
-                      loading="lazy"
-                      decoding="async"
-                      className="w-full h-full object-contain mix-blend-multiply dark:mix-blend-normal"
-                    />
-                  </Link>
-                  <div
-                    className="flex text-yellow-400 mb-2"
-                    role="img"
-                    aria-label={`Rating: ${product.rating || 5} out of 5 stars`}
-                  >
-                    {[...Array(5)].map((_, i) => (
-                      <Star
-                        key={i}
-                        className="w-4 h-4 fill-current"
-                        aria-hidden="true"
-                      />
-                    ))}
-                  </div>
-                  <h4 className="font-bold text-gray-900 dark:text-white mb-2 line-clamp-2 hover:text-green-600 transition-colors">
-                    <Link to={`/product/${product._id}`}>{product.name}</Link>
-                  </h4>
-                  <div className="mt-auto">
-                    <p className="text-xl font-black text-gray-900 dark:text-white mb-4">
-                      ${product.price.toFixed(2)}
-                    </p>
-                    <button
-                      aria-label={`Add one box of ${product.name} to cart`}
-                      onClick={() => handleAddToCart(product)}
-                      disabled={addingProductId === product._id}
-                      className="w-full bg-green-700 hover:bg-green-800 disabled:bg-green-500 disabled:cursor-not-allowed text-white font-bold py-2.5 rounded-lg flex items-center justify-center gap-2 transition-transform active:scale-95"
-                    >
-                      {addingProductId === product._id ? (
-                        <Loader2
-                          className="w-5 h-5 animate-spin"
-                          aria-hidden="true"
-                        />
-                      ) : (
-                        <ShoppingCart className="w-5 h-5" aria-hidden="true" />
-                      )}
-                      {addingProductId === product._id
-                        ? "Adding..."
-                        : "Add To Cart"}
-                    </button>
-                  </div>
-                </div>
-              ))
-            ) : (
-              <div className="col-span-full text-center text-gray-500 py-10 text-xl font-semibold">
-                No products found.
-              </div>
-            )}
-          </div>
-        )}
-      </div>
-
-      <div className="w-full px-4 sm:px-8 lg:px-12 py-12">
-        <div className="grid grid-cols-1 md:grid-cols-2 gap-6">
-          <img
-            src={OfferOne}
-            alt="Special Offer 1"
-            width="800"
-            height="400"
-            loading="lazy"
-            decoding="async"
-            className="w-full h-auto rounded-2xl shadow-md object-cover aspect-[2/1]"
-          />
-          <img
-            src={OfferTwo}
-            alt="Special Offer 2"
-            width="800"
-            height="400"
-            loading="lazy"
-            decoding="async"
-            className="w-full h-auto rounded-2xl shadow-md object-cover aspect-[2/1]"
-          />
-        </div>
-      </div>
-
-      <div className="w-full bg-gray-100 dark:bg-[#1a1a1a] py-16 transition-colors duration-300">
-        <div className="w-full px-4 sm:px-8 lg:px-12">
-          <h3 className="text-3xl font-bold text-center text-gray-900 dark:text-white mb-12">
-            What Our Clients Say
-          </h3>
-          <div className="grid grid-cols-1 md:grid-cols-3 gap-8">
-            {[
-              {
-                img: Abdo,
-                name: "Abdelrahman Elaraby",
-                text: "Excellent service! The delivery was fast and the products are top quality.",
-                rating: 5,
-              },
-              {
-                img: Selim,
-                name: "Abdelrahman Selim",
-                text: "Very professional staff and great customer support. Highly recommended!",
-                rating: 5,
-              },
-              {
-                img: Three,
-                name: "Omar Khaled",
-                text: "Affordable prices and authentic products. Will order again for sure.",
-                rating: 4,
-              },
-            ].map((client, idx) => (
-              <div
-                key={idx}
-                className="bg-white dark:bg-[#252525] p-8 rounded-2xl text-center shadow-sm border border-gray-100 dark:border-gray-800 relative pt-14 mt-8"
-              >
-                <img
-                  src={client.img}
-                  alt={`Client ${client.name}`}
-                  loading="lazy"
-                  className="w-24 h-24 rounded-full object-cover absolute -top-12 left-1/2 transform -translate-x-1/2 border-4 border-white dark:border-[#252525] shadow-md"
-                />
-                <h4 className="font-bold text-lg text-gray-900 dark:text-white mb-3">
-                  {client.name}
-                </h4>
-                <p className="text-gray-600 dark:text-gray-400 italic mb-4 leading-relaxed">
-                  "{client.text}"
-                </p>
-                <div
-                  className="flex justify-center text-yellow-400"
-                  role="img"
-                  aria-label={`Rating: ${client.rating} out of 5 stars`}
-                >
-                  {[...Array(5)].map((_, i) => (
-                    <Star
-                      key={i}
-                      className={`w-5 h-5 ${i < client.rating ? "fill-current" : "text-gray-300 dark:text-gray-600"}`}
-                      aria-hidden="true"
-                    />
-                  ))}
-                </div>
-              </div>
-            ))}
-          </div>
-        </div>
-      </div>
-
-      <div className="max-w-5xl mx-auto px-4 sm:px-8 lg:px-12 py-20">
-        <h3 className="text-3xl font-bold text-center text-gray-900 dark:text-white mb-10">
-          Frequently Asked Questions
-        </h3>
-        <div className="space-y-4">
-          {faqs.map((faq, index) => (
-            <div
-              key={index}
-              className={`border border-gray-200 dark:border-gray-700 rounded-xl overflow-hidden transition-all duration-300 ${activeIndex === index ? "bg-green-50 dark:bg-green-900/10 border-green-500" : "bg-white dark:bg-[#1e1e1e]"}`}
-            >
-              <button
-                aria-expanded={activeIndex === index}
-                onClick={() => toggleFAQ(index)}
-                className="w-full px-6 py-4 flex justify-between items-center focus:outline-none"
-              >
-                <span
-                  className={`font-semibold text-left ${activeIndex === index ? "text-green-700 dark:text-green-400" : "text-gray-900 dark:text-white"}`}
-                >
-                  {faq.question}
-                </span>
-                <span
-                  className={`text-2xl font-light transition-transform duration-300 ${activeIndex === index ? "text-green-600 rotate-45" : "text-gray-400"}`}
-                  aria-hidden="true"
-                >
-                  +
-                </span>
-              </button>
-              <div
-                className={`px-6 overflow-hidden transition-all duration-300 ease-in-out ${activeIndex === index ? "max-h-40 py-4 border-t border-green-100 dark:border-green-800" : "max-h-0"}`}
-              >
-                <p className="text-gray-600 dark:text-gray-400">{faq.answer}</p>
-              </div>
-            </div>
-          ))}
-        </div>
-      </div>
-
-      <style>{`
-        @keyframes float {
-          0%, 100% { transform: translateY(0) scale(1.05); }
-          50% { transform: translateY(-20px) scale(1.05); }
-        }
-      `}</style>
+      <Suspense fallback={<globalLoader/>}>
+        <PromoBanners />
+        <TopProducts />
+        <SpecialOffers />
+        <Testimonials />
+        <FAQAccordion />
+      </Suspense>
     </div>
   );
 }
